refactor: move derived stores to neuron.derived
