+++ frontend/src/lib/stores/neurons.store.ts
@@ -1,12 +1,5 @@
-import {
-  hasValidStake,
-  shouldDisplayRewardLossNotification,
-  sortNeuronsByStake,
-  sortNeuronsByVotingPowerRefreshedTimeout,
-} from "$lib/utils/neuron.utils";
 import type { NeuronInfo } from "@dfinity/nns";
-import { nonNullish } from "@dfinity/utils";
-import { derived, writable, type Readable } from "svelte/store";
+import { writable } from "svelte/store";
 
 export interface NeuronsStore {
   neurons?: NeuronInfo[];
@@ -88,31 +81,3 @@ const initNeuronsStore = () => {
 };
 
 export const neuronsStore = initNeuronsStore();
-
-export const definedNeuronsStore: Readable<NeuronInfo[]> = derived(
-  neuronsStore,
-  ($neuronsStore) => $neuronsStore.neurons?.filter(hasValidStake) || []
-);
-
-export const sortedNeuronStore: Readable<NeuronInfo[]> = derived(
-  definedNeuronsStore,
-  (initializedNeuronsStore) => sortNeuronsByStake(initializedNeuronsStore)
-);
-
-export const soonLosingRewardNeuronsStore: Readable<NeuronInfo[]> = derived(
-  definedNeuronsStore,
-  ($definedNeuronsStore) =>
-    sortNeuronsByVotingPowerRefreshedTimeout(
-      $definedNeuronsStore.filter(shouldDisplayRewardLossNotification)
-    )
-);
-
-export const neuronAccountsStore: Readable<Set<string>> = derived(
-  neuronsStore,
-  ($neuronsStore) =>
-    new Set(
-      $neuronsStore.neurons
-        ?.map(({ fullNeuron }) => fullNeuron?.accountIdentifier)
-        .filter(nonNullish) || []
-    )
-);
